% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{predict.dbtmb}
\alias{predict.dbtmb}
\title{Predict from fitted dbtmb}
\usage{
\method{predict}{dbtmb}(
  fit,
  par,
  type = c("eversex"),
  cutoff = 24,
  long = TRUE,
  quick = FALSE,
  ...
)
}
\arguments{
\item{fit}{fitted dbtmb class}

\item{par}{parameter set to simulate}

\item{cutoff}{vector, can be length one}

\item{long}{long form}

\item{quick}{array form}
}
\description{
Predict from fitted dbtmb
}
